requirements.txt
Flask==2.3.3
Pillow==10.0.1
selenium==4.15.2
webdriver-manager==4.0.1
pytest==7.4.3
pytest-selenium==4.0.1
APScheduler==3.10.4
Flask-WTF==1.2.1
python-dotenv==1.0.1
PyJWT==2.8.0
